requirements.txt
# ARK Agent AGI Requirements

# Core Dependencies
google-generativeai>=0.3.0
pydantic>=2.0.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
numpy>=1.24.0
requests>=2.31.0

# Environment Management
python-dotenv>=1.0.0

# Built-in Tools Dependencies
Pillow>=10.0.0  # Image Processing Tool
fpdf2>=2.7.0    # PDF Generation Tool

# Optional API Integration (uncomment if needed)
# ARK Agent AGI Requirements

# Core Dependencies
google-generativeai>=0.3.0
pydantic>=2.0.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
numpy>=1.24.0
requests>=2.31.0

# Environment Management
python-dotenv>=1.0.0

# Built-in Tools Dependencies
Pillow>=10.0.0  # Image Processing Tool
fpdf2>=2.7.0    # PDF Generation Tool

# Optional API Integration (uncomment if needed)
# google-cloud-translate>=3.11.0  # For Translation Tool with Google API

# Development Dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
types-requests
fastapi==0.95.2
uvicorn==0.22.0
gunicorn==20.1.0
google-cloud-storage==2.10.0
joblib==1.3.2
scikit-learn==1.3.2
sentence-transformers==2.2.2
faiss-cpu==1.7.4.post2
pandas==2.2.2
sqlalchemy==2.1.0
sqlite-utils==3.40
mypy>=1.5.0
black>=23.0.0
isort>=5.12.0
types-requests>=2.31.0
